@@ -31,7 +31,7 @@ namespace NineChronicles.Headless
                 services.AddSingleton(provider => service);
                 services.AddSingleton(provider => service.Swarm);
                 services.AddSingleton(provider => service.BlockChain);
-                services.AddSingleton(provider => properties.Libplanet);
+                services.AddSingleton(provider => properties.Libplanet!);
                 services.AddSingleton(provider =>
                 {
                     return new ActionEvaluationPublisher(
